pyproject.toml
[project]
name = "ml-launchpad"
version = "0.1.0"
description = "AutoML consulting starter — built for Claude Code"
requires-python = ">=3.11"
dependencies = [
    "autogluon>=1.1",
    "scikit-learn>=1.4",
    "duckdb>=0.10",
    "mlflow>=2.12",
    "fastapi>=0.111",
    "uvicorn>=0.29",
    "streamlit>=1.35",
    "pandas>=2.2",
    "pyarrow>=16",
    "pyyaml>=6.0",
    "httpx>=0.27",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2",
    "pytest-cov>=5.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
